pyproject.toml
[project]
name = "mentorlib-sme"
version = "0.1.0"
description = "Backend of Mentorlib"
authors = [
    {name = "Kamel OURZIK"},
    {name = "Rayan CHOUCHANE"}]

readme = "README.md"

requires-python = ">3.7"

dependencies = [
  "flask", 
  "flask_restful",
  "flask_cors", 
  "flask_mail",
  "flask_jwt_extended",
  "flask_rq2",
  "flask_caching",
  "psycopg2", 
  "config", 
  "requests",  
  "python-dotenv" , 
  "bcrypt",
  "flask_sqlalchemy",
  "PyJWT",
  "flask_migrate",
  "flask_expects_json",
  "setuptools",
  "marshmallow_sqlalchemy",
  "alembic",
  "pydantic_settings",
  "pydantic",
  ]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "bandit[toml]==1.7.4",
    "black==22.10.0",
    "coverage[toml]==7.2.6",
    "pycodestyle==2.10.0",
    "pylint>=2.15.16",]


[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = ''

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
python_paths = "./"
testpaths = "tests/"
addopts = "-vv"

[pycodestyle]
count = true
max-line-length = 120
statistics = true

[tool.pylint.main]
ignore-paths = []

[tool.pylint.format]
max-line-length=120

[tool.pylint.reports]
msg-template="{path}:{line}: [{msg_id}, {obj}] {msg} ({symbol})"
output-format="parseable"

[tool.pylint.'messages control']
disable=[]

[tool.bandit]
targets = ["app"]
exclude_dirs = []

[tool.packages]
include = [{from="src", include=["mentorlib_sme"]}]